Expose bliss hydra executable in shipped package
